refactor: Update Photo component with new image and animation
The Photo component in the UI folder has been updated to use a new image file and animation. The image source has been changed to "/assets/photo2.png" and the opacity of the image has been adjusted to 80%. Additionally, a new SVG circle element has been added with a stroke color of "#6f00ff" and an animation that changes the stroke dash array and rotates the circle. This update improves the visual appearance of the component.