refactor: Update contact form inputs with dynamic values
This commit updates the contact form inputs on the contact page to use dynamic values from the `contactData` object. The `contactData.formData` object is now used to set the `type` and `placeholder` attributes of the input fields. This change allows for more flexibility and easier maintenance of the form inputs.